GHC.Num.BigNat

Description

Multi-precision natural

Documentation

type BigNat# = WordArray#

A BigNat

Represented as an array of limbs (Word#) stored in little-endian order (Word# themselves use machine order).

Invariant (canonical representation): higher Word# is non-zero.

As a consequence, zero is represented with a WordArray# whose size is 0.

data BigNat

A lifted BigNat

Represented as an array of limbs (Word#) stored in little-endian order (Word# themselves use machine order).

Invariant (canonical representation): higher Word# is non-zero.

As a consequence, zero is represented with a WordArray# whose size is 0.

Constructors

BN#

Fields

unBigNat :: BigNat#

bigNatCheck# :: BigNat# -> Bool#

Check that the BigNat is valid

bigNatCheck :: BigNat# -> Bool

Check that the BigNat is valid

bigNatSize :: BigNat# -> Word

Number of words in the BigNat

bigNatSize# :: BigNat# -> Int#

Number of words in the BigNat

bigNatZero :: BigNat

bigNatOne :: BigNat

bigNatZero# :: (# #) -> BigNat#

BigNat Zero

bigNatOne# :: (# #) -> BigNat#

BigNat one

raiseDivZero_BigNat :: (# #) -> BigNat#

bigNatIsZero :: BigNat# -> Bool

Indicate if a bigNat is zero

bigNatIsZero# :: BigNat# -> Bool#

Indicate if a bigNat is zero

bigNatIsOne :: BigNat# -> Bool

Indicate if a bigNat is one

bigNatIsOne# :: BigNat# -> Bool#

Indicate if a bigNat is one

bigNatIsTwo :: BigNat# -> Bool

Indicate if a bigNat is two

bigNatIsTwo# :: BigNat# -> Bool#

Indicate if a bigNat is two

bigNatIsPowerOf2# :: BigNat# -> (# (# #) | Word# #)

Indicate if the value is a power of two and which one

bigNatIndex# :: BigNat# -> Int# -> Word#

Return the Word# at the given index

bigNatIndex :: BigNat# -> Int# -> Word

Return the Word# at the given index

bigNatFromWord :: Word -> BigNat#

Create a BigNat from a Word

bigNatFromWord# :: Word# -> BigNat#

Create a BigNat from a Word

bigNatFromWordList :: [Word] -> BigNat#

Convert a list of non-zero Words (most-significant first) into a BigNat

bigNatFromWordList# :: [Word] -> WordArray#

Convert a list of non-zero Words (most-significant first) into a BigNat

bigNatFromAbsInt# :: Int# -> BigNat#

Return the absolute value of the Int# in a BigNat

bigNatFromWordListUnsafe :: [Word] -> BigNat#

Convert a list of non-zero Words (most-significant first) into a BigNat. Don't remove most-significant zero words

bigNatToWordList :: BigNat# -> [Word]

Convert a BigNat into a list of non-zero Words (most-significant first)

bigNatFromWord2# :: Word# -> Word# -> BigNat#

Convert two Word# (most-significant first) into a BigNat

bigNatToWord# :: BigNat# -> Word#

Convert a BigNat into a Word#

bigNatToWordMaybe# :: BigNat# -> (# (# #) | Word# #)

Convert a BigNat into a Word# if it fits

bigNatToWord :: BigNat# -> Word

Convert a BigNat into a Word

bigNatToInt# :: BigNat# -> Int#

Convert a BigNat into a Int#

bigNatToInt :: BigNat# -> Int

Convert a BigNat into a Int

bigNatFromWord64# :: Word64# -> BigNat#

Convert a Word64# into a BigNat on 64-bit architectures

bigNatToWord64# :: BigNat# -> Word64#

Convert a BigNat into a Word64# on 64-bit architectures

bigNatEncodeDouble# :: BigNat# -> Int# -> Double#

Encode (# BigNat mantissa, Int# exponent #) into a Double#

bigNatGtWord# :: BigNat# -> Word# -> Bool#

Test if a BigNat is greater than a Word

bigNatEqWord# :: BigNat# -> Word# -> Bool#

Test if a BigNat is equal to a Word

bigNatGtWord :: BigNat# -> Word -> Bool

Test if a BigNat is greater than a Word

bigNatLeWord# :: BigNat# -> Word# -> Bool#

Test if a BigNat is lower than or equal to a Word

bigNatLeWord :: BigNat# -> Word -> Bool

Test if a BigNat is lower than or equal to a Word

bigNatEq# :: BigNat# -> BigNat# -> Bool#

Equality test for BigNat

bigNatEq :: BigNat# -> BigNat# -> Bool

Equality test for BigNat

bigNatNe# :: BigNat# -> BigNat# -> Bool#

Inequality test for BigNat

bigNatNe :: BigNat# -> BigNat# -> Bool

Equality test for BigNat

bigNatCompareWord# :: BigNat# -> Word# -> Ordering

Compare a BigNat and a Word#

bigNatCompareWord :: BigNat# -> Word -> Ordering

Compare a BigNat and a Word

bigNatCompare :: BigNat# -> BigNat# -> Ordering

Compare two BigNat

bigNatLt# :: BigNat# -> BigNat# -> Bool#

Predicate: a < b

bigNatLt :: BigNat# -> BigNat# -> Bool

Predicate: a < b

bigNatLe# :: BigNat# -> BigNat# -> Bool#

Predicate: a <= b

bigNatLe :: BigNat# -> BigNat# -> Bool

Predicate: a <= b

bigNatGt# :: BigNat# -> BigNat# -> Bool#

Predicate: a > b

bigNatGt :: BigNat# -> BigNat# -> Bool

Predicate: a > b

bigNatGe# :: BigNat# -> BigNat# -> Bool#

Predicate: a >= b

bigNatGe :: BigNat# -> BigNat# -> Bool

Predicate: a >= b

bigNatAddWord# :: BigNat# -> Word# -> BigNat#

Add a bigNat and a Word#

bigNatAddWord :: BigNat# -> Word -> BigNat#

Add a bigNat and a Word

bigNatAdd :: BigNat# -> BigNat# -> BigNat#

Add two bigNats

bigNatMulWord# :: BigNat# -> Word# -> BigNat#

Multiply a BigNat by a Word#

bigNatMulWord :: BigNat# -> Word -> BigNat#

Multiply a BigNAt by a Word

bigNatSqr :: BigNat# -> BigNat#

Square a BigNat

bigNatMul :: BigNat# -> BigNat# -> BigNat#

Multiplication (classical algorithm)

bigNatSubWordUnsafe# :: BigNat# -> Word# -> BigNat#

Subtract a Word# from a BigNat

The BigNat must be bigger than the Word#.

bigNatSubWordUnsafe :: BigNat# -> Word -> BigNat#

Subtract a Word# from a BigNat

The BigNat must be bigger than the Word#.

bigNatSubWord# :: BigNat# -> Word# -> (# (# #) | BigNat# #)

Subtract a Word# from a BigNat

bigNatSubUnsafe :: BigNat# -> BigNat# -> BigNat#

Subtract two BigNat (don't check if a >= b)

bigNatSub :: BigNat# -> BigNat# -> (# (# #) | BigNat# #)

Subtract two BigNat

bigNatQuotWord# :: BigNat# -> Word# -> BigNat#

Divide a BigNat by a Word, return the quotient

Require: b /= 0

bigNatQuotWord :: BigNat# -> Word -> BigNat#

Divide a BigNat by a Word, return the quotient

Require: b /= 0

bigNatRemWord# :: BigNat# -> Word# -> Word#

Divide a BigNat by a Word, return the remainder

Require: b /= 0

bigNatRemWord :: BigNat# -> Word -> Word

Divide a BigNat by a Word, return the remainder

Require: b /= 0

bigNatQuotRemWord# :: BigNat# -> Word# -> (# BigNat#, Word# #)

QuotRem a BigNat by a Word

Require: b /= 0

bigNatQuotRem# :: BigNat# -> BigNat# -> (# BigNat#, BigNat# #)

BigNat division returning (quotient,remainder)

bigNatQuot :: BigNat# -> BigNat# -> BigNat#

BigNat division returning quotient

bigNatRem :: BigNat# -> BigNat# -> BigNat#

BigNat division returning remainder

gcdWord# :: Word# -> Word# -> Word#

Greatest common divisor between two Word#

gcdWord :: Word -> Word -> Word

Greatest common divisor between two Word

gcdInt# :: Int# -> Int# -> Int#

Greatest common divisor between two Int#

Warning: result may become negative if (at least) one argument is minBound

gcdInt :: Int -> Int -> Int

Greatest common divisor between two Int

Warning: result may become negative if (at least) one argument is minBound

bigNatGcd :: BigNat# -> BigNat# -> BigNat#

Greatest common divisor

bigNatGcdWord# :: BigNat# -> Word# -> Word#

Greatest common divisor

bigNatLcm :: BigNat# -> BigNat# -> BigNat#

Least common multiple

bigNatLcmWord# :: BigNat# -> Word# -> BigNat#

Least common multiple with a Word#

bigNatLcmWordWord# :: Word# -> Word# -> BigNat#

Least common multiple between two Word#

bigNatOr :: BigNat# -> BigNat# -> BigNat#

Bitwise OR

bigNatOrWord# :: BigNat# -> Word# -> BigNat#

Bitwise OR with Word#

bigNatAnd :: BigNat# -> BigNat# -> BigNat#

Bitwise AND

bigNatAndNot :: BigNat# -> BigNat# -> BigNat#

Bitwise ANDNOT

bigNatAndWord# :: BigNat# -> Word# -> BigNat#

Bitwise AND with Word#

bigNatAndNotWord# :: BigNat# -> Word# -> BigNat#

Bitwise ANDNOT with Word#

bigNatAndInt# :: BigNat# -> Int# -> BigNat#

Bitwise AND with Int#

bigNatXor :: BigNat# -> BigNat# -> BigNat#

Bitwise XOR

bigNatXorWord# :: BigNat# -> Word# -> BigNat#

Bitwise XOR with Word#

bigNatPopCount :: BigNat# -> Word

PopCount for BigNat

bigNatPopCount# :: BigNat# -> Word#

PopCount for BigNat

bigNatShiftR# :: BigNat# -> Word# -> BigNat#

Bit shift right

bigNatShiftRNeg# :: BigNat# -> Word# -> BigNat#

Bit shift right (two's complement)

bigNatShiftR :: BigNat# -> Word -> BigNat#

Bit shift right

bigNatShiftL :: BigNat# -> Word -> BigNat#

Bit shift left

bigNatShiftL# :: BigNat# -> Word# -> BigNat#

Bit shift left

bigNatTestBit# :: BigNat# -> Word# -> Bool#

BigNat bit test

bigNatTestBit :: BigNat# -> Word -> Bool

BigNat bit test

bigNatBit# :: Word# -> BigNat#

Return a BigNat whose bit i is the only one set.

Specialized version of `bigNatShiftL (bigNatFromWord# 1##)`

bigNatBit :: Word -> BigNat#

Return a BigNat whose bit i is the only one set.

Specialized version of `bigNatShiftL (bigNatFromWord# 1##)`

bigNatClearBit# :: BigNat# -> Word# -> BigNat#

BigNat clear bit

bigNatSetBit# :: BigNat# -> Word# -> BigNat#

BigNat set bit

bigNatComplementBit# :: BigNat# -> Word# -> BigNat#

Reverse the given bit

bigNatLog2# :: BigNat# -> Word#

Base 2 logarithm

bigNatLog2 :: BigNat# -> Word

Base 2 logarithm

bigNatLogBase# :: BigNat# -> BigNat# -> Word#

Logarithm for an arbitrary base

bigNatLogBase :: BigNat# -> BigNat# -> Word

Logarithm for an arbitrary base

bigNatLogBaseWord# :: Word# -> BigNat# -> Word#

Logarithm for an arbitrary base

bigNatLogBaseWord :: Word -> BigNat# -> Word

Logarithm for an arbitrary base

bigNatSizeInBase# :: Word# -> BigNat# -> Word#

Compute the number of digits of the BigNat in the given base.

base must be > 1

bigNatSizeInBase :: Word -> BigNat# -> Word

Compute the number of digits of the BigNat in the given base.

base must be > 1

powModWord# :: Word# -> Word# -> Word# -> Word#

bigNatPowModWord# :: BigNat# -> BigNat# -> Word# -> Word#

"bigNatPowModWord# b e m" computes base b raised to exponent e modulo m.

bigNatPowMod :: BigNat# -> BigNat# -> BigNat# -> BigNat#

"bigNatPowMod b e m" computes base b raised to exponent e modulo m.

bigNatCtz# :: BigNat# -> Word#

Return count of trailing zero bits

Return 0 for zero BigNat

bigNatCtz :: BigNat# -> Word

Return count of trailing zero bits

Return 0 for zero BigNat

bigNatCtzWord# :: BigNat# -> Word#

Return count of trailing zero words

Return 0 for zero BigNat

bigNatCtzWord :: BigNat# -> Word

Return count of trailing zero words

Return 0 for zero BigNat

bigNatToAddrLE# :: BigNat# -> Addr# -> State# s -> (# State# s, Word# #)

Write a BigNat in base-256 little-endian representation and return the number of bytes written.

Use "bigNatSizeInBase 256# i" to compute the exact number of bytes written in advance. In case of i == 0, the function will write and report zero bytes written.

bigNatToAddrBE# :: BigNat# -> Addr# -> State# s -> (# State# s, Word# #)

Write a BigNat in base-256 big-endian representation and return the number of bytes written.

Use "bigNatSizeInBase 256# i" to compute the exact number of bytes written in advance. In case of i == 0, the function will write and report zero bytes written.

bigNatToAddr# :: BigNat# -> Addr# -> Bool# -> State# s -> (# State# s, Word# #)

Write a BigNat in base-256 representation and return the number of bytes written.

The endianness is selected with the Bool# parameter: most significant byte first (big-endian) if 1# or least significant byte first (little-endian) if 0#.

Use "bigNatSizeInBase 256# i" to compute the exact number of bytes written in advance. In case of i == 0, the function will write and report zero bytes written.

bigNatToAddr :: BigNat# -> Addr# -> Bool# -> IO Word

Write a BigNat in base-256 representation and return the number of bytes written.

The endianness is selected with the Bool# parameter: most significant byte first (big-endian) if 1# or least significant byte first (little-endian) if 0#.

Use "bigNatSizeInBase 256# i" to compute the exact number of bytes written in advance. In case of i == 0, the function will write and report zero bytes written.

bigNatFromAddrLE# :: Word# -> Addr# -> State# s -> (# State# s, BigNat# #)

Read a BigNat in base-256 little-endian representation from an Addr#.

The size is given in bytes.

Higher limbs equal to 0 are automatically trimmed.

bigNatFromAddrBE# :: Word# -> Addr# -> State# s -> (# State# s, BigNat# #)

Read a BigNat in base-256 big-endian representation from an Addr#.

The size is given in bytes.

Null higher limbs are automatically trimmed.

bigNatFromAddr# :: Word# -> Addr# -> Bool# -> State# s -> (# State# s, BigNat# #)

Read a BigNat in base-256 representation from an Addr#.

The size is given in bytes.

The endianness is selected with the Bool# parameter: most significant byte first (big-endian) if 1# or least significant byte first (little-endian) if 0#.

Null higher limbs are automatically trimmed.

bigNatToMutableByteArrayLE# :: BigNat# -> MutableByteArray# s -> Word# -> State# s -> (# State# s, Word# #)

Write a BigNat in base-256 little-endian representation and return the number of bytes written.

Use "bigNatSizeInBase 256# i" to compute the exact number of bytes written in advance. In case of i == 0, the function will write and report zero bytes written.

bigNatToMutableByteArrayBE# :: BigNat# -> MutableByteArray# s -> Word# -> State# s -> (# State# s, Word# #)

Write a BigNat in base-256 big-endian representation and return the number of bytes written.

Use "bigNatSizeInBase 256# i" to compute the exact number of bytes written in advance. In case of i == 0, the function will write and report zero bytes written.

bigNatToMutableByteArray# :: BigNat# -> MutableByteArray# s -> Word# -> Bool# -> State# s -> (# State# s, Word# #)

Write a BigNat in base-256 representation and return the number of bytes written.

The endianness is selected with the Bool# parameter: most significant byte first (big-endian) if 1# or least significant byte first (little-endian) if 0#.

Use "bigNatSizeInBase 256# i" to compute the exact number of bytes written in advance. In case of i == 0, the function will write and report zero bytes written.

bigNatFromByteArrayLE# :: Word# -> ByteArray# -> Word# -> State# s -> (# State# s, BigNat# #)

Read a BigNat in base-256 little-endian representation from a ByteArray#.

The size is given in bytes.

Null higher limbs are automatically trimmed.

bigNatFromByteArrayBE# :: Word# -> ByteArray# -> Word# -> State# s -> (# State# s, BigNat# #)

Read a BigNat in base-256 big-endian representation from a ByteArray#.

The size is given in bytes.

Null higher limbs are automatically trimmed.

bigNatFromByteArray# :: Word# -> ByteArray# -> Word# -> Bool# -> State# s -> (# State# s, BigNat# #)

Read a BigNat in base-256 representation from a ByteArray#.

The size is given in bytes.

The endianness is selected with the Bool# parameter: most significant byte first (big-endian) if 1# or least significant byte first (little-endian) if 0#.

Null higher limbs are automatically trimmed.

bigNatFromWordArray# :: WordArray# -> Word# -> BigNat#

Create a BigNat# from a WordArray# containing n limbs in least-significant-first order.

If possible WordArray#, will be used directly (i.e. shared without cloning the WordArray# into a newly allocated one)

bigNatFromWordArray :: WordArray# -> Word# -> BigNat

Create a BigNat from a WordArray# containing n limbs in least-significant-first order.

If possible WordArray#, will be used directly (i.e. shared without cloning the WordArray# into a newly allocated one)