GHC.Num.Integer

Description

The Integer type.

Documentation

data Integer

Arbitrary precision integers. In contrast with fixed-size integral types such as Int, the Integer type represents the entire infinite range of integers.

Integers are stored in a kind of sign-magnitude form, hence do not expect two's complement form when using bit operations.

If the value is small (fit into an Int), IS constructor is used. Otherwise Integer and IN constructors are used to store a BigNat representing respectively the positive or the negative value magnitude.

Invariant: Integer and IN are used iff value doesn't fit in IS

Constructors

IS !Int#
iff value in [minBound::Int, maxBound::Int] range

IP !BigNat#
iff value in ]maxBound::Int, +inf[ range

IN !BigNat#
iff value in ]-inf, minBound::Int[ range

integerCheck# :: Integer -> Bool#

Check Integer invariants

integerCheck :: Integer -> Bool

Check Integer invariants

integerZero :: Integer

Integer Zero

integerOne :: Integer

Integer One

integerFromBigNat# :: BigNat# -> Integer

Create a positive Integer from a BigNat

integerFromBigNatNeg# :: BigNat# -> Integer

Create a negative Integer from a BigNat

integerFromBigNatSign# :: Int# -> BigNat# -> Integer

Create an Integer from a sign-bit and a BigNat

integerToBigNatSign# :: Integer -> (# Int#, BigNat# #)

Convert an Integer into a sign-bit and a BigNat

integerToBigNatClamp# :: Integer -> BigNat#

Convert an Integer into a BigNat.

Return 0 for negative Integers.

integerFromInt# :: Int# -> Integer

Create an Integer from an Int#

integerFromInt :: Int -> Integer

Create an Integer from an Int

integerToInt# :: Integer -> Int#

Truncates Integer to least-significant Int#

integerToInt :: Integer -> Int

Truncates Integer to least-significant Int#

integerFromWord# :: Word# -> Integer

Convert a Word# into an Integer

integerFromWord :: Word -> Integer

Convert a Word into an Integer

integerFromWordNeg# :: Word# -> Integer

Create a negative Integer with the given Word magnitude

integerFromWordSign# :: Int# -> Word# -> Integer

Create an Integer from a sign and a Word magnitude

integerToWord# :: Integer -> Word#

Truncate an Integer into a Word

integerToWord :: Integer -> Word

Truncate an Integer into a Word

integerFromNatural :: Natural -> Integer

Convert a Natural into an Integer

integerFromWordList :: Bool -> [Word] -> Integer

Convert a list of Word into an Integer

integerToNaturalClamp :: Integer -> Natural

Convert an Integer into a Natural

Return 0 for negative Integers.

integerToNatural :: Integer -> Natural

Convert an Integer into a Natural

Return absolute value

integerToNaturalThrow :: Integer -> Natural

Convert an Integer into a Natural

Throw an Underflow exception if input is negative.

integerIsNegative# :: Integer -> Bool#

Negative predicate

integerIsNegative :: Integer -> Bool

Negative predicate

integerIsZero :: Integer -> Bool

Zero predicate

integerIsOne :: Integer -> Bool

One predicate

integerNe :: Integer -> Integer -> Bool

Not-equal predicate.

integerEq :: Integer -> Integer -> Bool

Equal predicate.

integerLe :: Integer -> Integer -> Bool

Lower-or-equal predicate.

integerLt :: Integer -> Integer -> Bool

Lower predicate.

integerGt :: Integer -> Integer -> Bool

Greater predicate.

integerGe :: Integer -> Integer -> Bool

Greater-or-equal predicate.

integerEq# :: Integer -> Integer -> Bool#

Equal predicate.

integerNe# :: Integer -> Integer -> Bool#

Not-equal predicate.

integerGt# :: Integer -> Integer -> Bool#

Greater predicate.

integerLe# :: Integer -> Integer -> Bool#

Lower-or-equal predicate.

integerLt# :: Integer -> Integer -> Bool#

Lower predicate.

integerGe# :: Integer -> Integer -> Bool#

Greater-or-equal predicate.

integerCompare :: Integer -> Integer -> Ordering

Compare two Integer

integerSub :: Integer -> Integer -> Integer

Subtract one Integer from another.

integerAdd :: Integer -> Integer -> Integer

Add two Integers

integerMul :: Integer -> Integer -> Integer

Multiply two Integers

integerNegate :: Integer -> Integer

Negate Integer.

One edge-case issue to take into account is that Int's range is not symmetric around 0. I.e. minBound+maxBound = -1

IP is used iff n > maxBound::Int IN is used iff n < minBound::Int

integerAbs :: Integer -> Integer

Compute absolute value of an Integer

integerSignum :: Integer -> Integer

Return -1, 0, and 1 depending on whether argument is negative, zero, or positive, respectively

integerSignum# :: Integer -> Int#

Return -1#, 0#, and 1# depending on whether argument is negative, zero, or positive, respectively

integerPopCount# :: Integer -> Int#

Count number of set bits. For negative arguments returns the negated population count of the absolute value.

integerBit# :: Word# -> Integer

Positive Integer for which only n-th bit is set

integerBit :: Word -> Integer

Integer for which only n-th bit is set

integerTestBit# :: Integer -> Word# -> Bool#

Test if n-th bit is set.

Fake 2's complement for negative values (might be slow)

integerTestBit :: Integer -> Word -> Bool

Test if n-th bit is set. For negative Integers it tests the n-th bit of the negated argument.

Fake 2's complement for negative values (might be slow)

integerShiftR# :: Integer -> Word# -> Integer

Shift-right operation

Fake 2's complement for negative values (might be slow)

integerShiftR :: Integer -> Word -> Integer

Shift-right operation

Fake 2's complement for negative values (might be slow)

integerShiftL# :: Integer -> Word# -> Integer

Shift-left operation

integerShiftL :: Integer -> Word -> Integer

Shift-left operation

Remember that bits are stored in sign-magnitude form, hence the behavior of negative Integers is different from negative Int's behavior.

integerOr :: Integer -> Integer -> Integer

Bitwise OR operation

Fake 2's complement for negative values (might be slow)

integerXor :: Integer -> Integer -> Integer

Bitwise XOR operation

Fake 2's complement for negative values (might be slow)

integerAnd :: Integer -> Integer -> Integer

Bitwise AND operation

Fake 2's complement for negative values (might be slow)

integerComplement :: Integer -> Integer

Binary complement of the

integerQuotRem# :: Integer -> Integer -> (# Integer, Integer #)

Simultaneous integerQuot and integerRem.

Divisor must be non-zero otherwise the GHC runtime will terminate with a division-by-zero fault.

integerQuotRem :: Integer -> Integer -> (Integer, Integer)

Simultaneous integerQuot and integerRem.

Divisor must be non-zero otherwise the GHC runtime will terminate with a division-by-zero fault.

integerQuot :: Integer -> Integer -> Integer

integerRem :: Integer -> Integer -> Integer

integerDivMod# :: Integer -> Integer -> (# Integer, Integer #)

Simultaneous integerDiv and integerMod.

Divisor must be non-zero otherwise the GHC runtime will terminate with a division-by-zero fault.

integerDivMod :: Integer -> Integer -> (Integer, Integer)

Simultaneous integerDiv and integerMod.

Divisor must be non-zero otherwise the GHC runtime will terminate with a division-by-zero fault.

integerDiv :: Integer -> Integer -> Integer

integerMod :: Integer -> Integer -> Integer

integerGcd :: Integer -> Integer -> Integer

Compute greatest common divisor.

integerLcm :: Integer -> Integer -> Integer

Compute least common multiple.

integerSqr :: Integer -> Integer

Square a Integer

integerLog2# :: Integer -> Word#

Base 2 logarithm (floor)

For numbers <= 0, return 0

integerLog2 :: Integer -> Word

Base 2 logarithm (floor)

For numbers <= 0, return 0

integerLogBaseWord# :: Word# -> Integer -> Word#

Logarithm (floor) for an arbitrary base

For numbers <= 0, return 0

integerLogBaseWord :: Word -> Integer -> Word

Logarithm (floor) for an arbitrary base

For numbers <= 0, return 0

integerLogBase# :: Integer -> Integer -> Word#

Logarithm (floor) for an arbitrary base

For numbers <= 0, return 0

integerLogBase :: Integer -> Integer -> Word

Logarithm (floor) for an arbitrary base

For numbers <= 0, return 0

integerIsPowerOf2# :: Integer -> (# (# #) | Word# #)

Indicate if the value is a power of two and which one

integerFromInt64# :: Int64# -> Integer

Convert an Int64# into an Integer

integerFromWord64# :: Word64# -> Integer

Convert a Word64# into an Integer

integerToInt64# :: Integer -> Int64#

Convert an Integer into an Int64#

integerToWord64# :: Integer -> Word64#

Convert an Integer into a Word64#

integerDecodeDouble# :: Double# -> (# Integer, Int# #)

Decode a Double# into (# Integer mantissa, Int# exponent #)

integerEncodeDouble# :: Integer -> Int# -> Double#

Encode (# Integer mantissa, Int# exponent #) into a Double#

integerEncodeDouble :: Integer -> Int -> Double

Encode (Integer mantissa, Int exponent) into a Double

integerEncodeFloat# :: Integer -> Int# -> Float#

Encode (# Integer mantissa, Int# exponent #) into a Float#

TODO: Not sure if it's worth to write Float optimized versions here

integerSizeInBase# :: Word# -> Integer -> Word#

Compute the number of digits of the Integer (without the sign) in the given base.

base must be > 1

integerToAddr# :: Integer -> Addr# -> Bool# -> State# s -> (# State# s, Word# #)

Write an Integer (without sign) to addr in base-256 representation and return the number of bytes written.

The endianness is selected with the Bool# parameter: write most significant byte first (big-endian) if 1# or least significant byte first (little-endian) if 0#.

integerToAddr :: Integer -> Addr# -> Bool# -> IO Word

Write an Integer (without sign) to addr in base-256 representation and return the number of bytes written.

The endianness is selected with the Bool# parameter: write most significant byte first (big-endian) if 1# or least significant byte first (little-endian) if 0#.

integerFromAddr# :: Word# -> Addr# -> Bool# -> State# s -> (# State# s, Integer #)

Read an Integer (without sign) in base-256 representation from an Addr#.

The size is given in bytes.

The endianness is selected with the Bool# parameter: most significant byte first (big-endian) if 1# or least significant byte first (little-endian) if 0#.

Null higher limbs are automatically trimed.

integerFromAddr :: Word# -> Addr# -> Bool# -> IO Integer

Read an Integer (without sign) in base-256 representation from an Addr#.

The size is given in bytes.

The endianness is selected with the Bool# parameter: most significant byte first (big-endian) if 1# or least significant byte first (little-endian) if 0#.

Null higher limbs are automatically trimed.

integerToMutableByteArray# :: Integer -> MutableByteArray# s -> Word# -> Bool# -> State# s -> (# State# s, Word# #)

Write an Integer (without sign) in base-256 representation and return the number of bytes written.

The endianness is selected with the Bool# parameter: most significant byte first (big-endian) if 1# or least significant byte first (little-endian) if 0#.

integerToMutableByteArray :: Integer -> MutableByteArray# RealWorld -> Word# -> Bool# -> IO Word

Write an Integer (without sign) in base-256 representation and return the number of bytes written.

The endianness is selected with the Bool# parameter: most significant byte first (big-endian) if 1# or least significant byte first (little-endian) if 0#.

integerFromByteArray# :: Word# -> ByteArray# -> Word# -> Bool# -> State# s -> (# State# s, Integer #)

Read an Integer (without sign) in base-256 representation from a ByteArray#.

The size is given in bytes.

The endianness is selected with the Bool# parameter: most significant byte first (big-endian) if 1# or least significant byte first (little-endian) if 0#.

Null higher limbs are automatically trimed.

integerFromByteArray :: Word# -> ByteArray# -> Word# -> Bool# -> Integer

Read an Integer (without sign) in base-256 representation from a ByteArray#.

The size is given in bytes.

The endianness is selected with the Bool# parameter: most significant byte first (big-endian) if 1# or least significant byte first (little-endian) if 0#.

Null higher limbs are automatically trimed.

integerGcde# :: Integer -> Integer -> (# Integer, Integer, Integer #)

Get the extended GCD of two integers.

`integerGcde# a b` returns (# g,x,y #) where * ax + by = g = |gcd a b|

integerGcde :: Integer -> Integer -> (Integer, Integer, Integer)

Get the extended GCD of two integers.

`integerGcde a b` returns (g,x,y) where * ax + by = g = |gcd a b|

integerRecipMod# :: Integer -> Natural -> (# Natural | () #)

Computes the modular inverse.

I.e. y = integerRecipMod# x m = x^(-1) mod m

with 0 < y < |m|

integerPowMod# :: Integer -> Integer -> Natural -> (# Natural | () #)

Computes the modular exponentiation.

I.e. y = integer_powmod b e m = b^e mod m

with 0 <= y < abs m

If e is negative, we use integerRecipMod# to try to find a modular multiplicative inverse (which may not exist).