GHC.Num.Primitives

Bool#

type Bool# = Int#

(&&#) :: Bool# -> Bool# -> Bool# infixr 3

(||#) :: Bool# -> Bool# -> Bool# infixr 2

notB# :: Bool# -> Bool#

Int#

testBitI# :: Int# -> Word# -> Bool#

minI# :: Int# -> Int# -> Int#

maxI# :: Int# -> Int# -> Int#

sgnI# :: Int# -> Int#

Branchless signum

absI# :: Int# -> Int#

Branchless abs

cmpI# :: Int# -> Int# -> Int#

Branchless comparison

intEncodeDouble# :: Int# -> Int# -> Double#

Encode (# Int# mantissa, Int# exponent #) into a Double#.

(provided by GHC's RTS)

popCntI# :: Int# -> Word#

Population count

Word#

andNot# :: Word# -> Word# -> Word#

cmpW# :: Word# -> Word# -> Ordering

bitW# :: Int# -> Word#

maxW# :: Word# -> Word# -> Word#

minW# :: Word# -> Word# -> Word#

testBitW# :: Word# -> Word# -> Bool#

shiftRW# :: Word# -> Word# -> Word#

Safe right shift for Word#

plusWord3# :: Word# -> Word# -> Word# -> (# Word#, Word# #)

Add 3 values together

plusWord12# :: Word# -> (# Word#, Word# #) -> (# Word#, Word# #)

(h,l) <- a + (hb,lb)

quotRemWord3# :: (# Word#, Word# #) -> Word# -> (# (# Word#, Word# #), Word# #)

2-by-1 large division

Requires: b0 /= 0 a1 >= b0 (not required, but if not q1=0)

wordFromAbsInt# :: Int# -> Word#

Return the absolute value of the Int# in a Word#

wordLog2# :: Word# -> Word#

Compute base-2 log of Word#

This is internally implemented as count-leading-zeros machine instruction.

wordLogBase# :: Word# -> Word# -> Word#

Logarithm for an arbitrary base

wordSizeInBase# :: Word# -> Word# -> Word#

wordIsPowerOf2# :: Word# -> (# (# #) | Word# #)

Indicate if the value is a power of two and which one

wordEncodeDouble# :: Word# -> Int# -> Double#

Encode (# Word# mantissa, Int# exponent #) into a Double#.

(provided by GHC's RTS)

wordReverseBits# :: Word# -> Word#

Reverse bits in a Word#

wordReverseBits32# :: Word# -> Word#

Reverse bits in the Word32 subwords composing a Word#

wordReverseBytes# :: Word# -> Word#

Reverse bytes in a Word#

Addr import/export

wordFromAddr# :: Word# -> Addr# -> Bool# -> State# s -> (# State# s, Word# #)

Read a Word from addr in base-256 representation.

@n is the number of bytes to read.

The endianness is selected with the Bool# parameter: write most significant byte first (big-endian) if 1# or least significant byte first (little-endian) if 0#.

wordFromAddrLE# :: Word# -> Addr# -> State# s -> (# State# s, Word# #)

Read a Word from addr in base-256 little-endian representation.

@n is the number of bytes to read.

wordFromAddrBE# :: Word# -> Addr# -> State# s -> (# State# s, Word# #)

Read a Word from addr in base-256 big-endian representation.

@n is the number of bytes to read.

wordToAddr# :: Word# -> Addr# -> Bool# -> State# s -> (# State# s, Word# #)

Write a Word to addr in base-256 representation and return the number of bytes written.

The endianness is selected with the Bool# parameter: write most significant byte first (big-endian) if 1# or least significant byte first (little-endian) if 0#.

wordToAddrLE# :: Word# -> Addr# -> State# s -> (# State# s, Word# #)

Write a Word to addr in base-256 little-endian representation and return the number of bytes written.

wordToAddrBE# :: Word# -> Addr# -> State# s -> (# State# s, Word# #)

Write a Word to addr in base-256 big-endian representation and return the number of bytes written.

wordWriteAddrLE# :: Word# -> Addr# -> State# s -> State# s

Write a full word with little-endian encoding

wordWriteAddrBE# :: Word# -> Addr# -> State# s -> State# s

Write a full word with little-endian encoding

ByteArray import/export

wordFromByteArray# :: Word# -> ByteArray# -> Word# -> Bool# -> Word#

Read a Word from ByteArray in base-256 representation.

@n is the number of bytes to read.

The endianness is selected with the Bool# parameter: write most significant byte first (big-endian) if 1# or least significant byte first (little-endian) if 0#.

wordFromByteArrayLE# :: Word# -> ByteArray# -> Word# -> Word#

Read a Word from ByteArray in base-256 little-endian representation.

@n is the number of bytes to read.

wordFromByteArrayBE# :: Word# -> ByteArray# -> Word# -> Word#

Read a Word from ByteArray in base-256 big-endian representation.

@n is the number of bytes to read.

wordToMutableByteArray# :: Word# -> MutableByteArray# s -> Word# -> Bool# -> State# s -> (# State# s, Word# #)

Write a Word to MutableByteArray in base-256 representation and return the number of bytes written.

The endianness is selected with the Bool# parameter: write most significant byte first (big-endian) if 1# or least significant byte first (little-endian) if 0#.

The offset is in bytes.

wordToMutableByteArrayLE# :: Word# -> MutableByteArray# s -> Word# -> State# s -> (# State# s, Word# #)

Write a Word to MutableByteArray in base-256 little-endian representation and return the number of bytes written.

The offset is in bytes.

wordToMutableByteArrayBE# :: Word# -> MutableByteArray# s -> Word# -> State# s -> (# State# s, Word# #)

Write a Word to MutableByteArray in base-256 big-endian representation and return the number of bytes written.

The offset is in bytes.

wordWriteMutableByteArrayLE# :: Word# -> MutableByteArray# s -> Word# -> State# s -> State# s

Write a full word with little-endian encoding

wordWriteMutableByteArrayBE# :: Word# -> MutableByteArray# s -> Word# -> State# s -> State# s

Write a full word with little-endian encoding

Exception

raiseUnderflow :: a

Raise underflowException

raiseUnderflow_Word# :: (# #) -> Word#

raiseDivZero :: a

Raise divZeroException

raiseDivZero_Word# :: (# #) -> Word#

unexpectedValue :: a

unexpectedValue_Int# :: (# #) -> Int#

unexpectedValue_Word# :: (# #) -> Word#

IO

ioWord# :: IO Word -> State# RealWorld -> (# State# RealWorld, Word# #)

ioInt# :: IO Int -> State# RealWorld -> (# State# RealWorld, Int# #)

ioVoid :: IO a -> State# RealWorld -> State# RealWorld

ioBool :: IO Bool -> State# RealWorld -> (# State# RealWorld, Bool# #)